Rename signature to be more generic (and less wrong); Make copyable accept children
