+++ src/scss/_solana.scss
@@ -10,7 +10,7 @@ code {
   color: $black;
 }
 
-.signature {
+.copyable {
   position: relative;
 
   &:hover {
